rename team file name
