+++ Web/starter_project_barbies/components/team/TeamMemberCard.tsx
